requirements.txt
# Core dependencies
python>=3.10

# API and Web Framework
fastapi==0.104.1
uvicorn[standard]==0.24.0
streamlit==1.28.1
pydantic==2.5.0
pydantic-settings==2.1.0

# LangChain and LangGraph
langchain==0.1.0
langchain-google-genai==1.0.0
langgraph==0.0.20
langchain-community==0.0.10

# Data processing
pandas==2.1.3
numpy==1.26.2
requests==2.31.0

# Vector Database
chromadb==0.4.18

# ML and MLOps
scikit-learn==1.3.2
mlflow==2.9.2
dvc==3.38.1
joblib==1.3.2

# Technical Analysis
ta==0.11.0

# Visualization
plotly==5.18.0

# Utilities
python-dotenv==1.0.0
pyyaml==6.0.1

# Testing
pytest==7.4.3
pytest-asyncio==0.21.1
black==23.11.0
ruff==0.1.6

